extracted generateHumanReadable() method out of ModInitializer implementing class
