+++ reference-mod/src/main/java/dev/mineblock11/fabric/referencemod/util/helper/TextUtil.java
@@ -1,14 +1,19 @@
 package dev.mineblock11.fabric.referencemod.util.helper;
 
+import dev.mineblock11.fabric.referencemod.MyMod;
 import net.minecraft.text.MutableText;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
+import net.minecraft.util.Identifier;
+import org.apache.commons.lang3.text.WordUtils;
+
 
 /**
  * <h2>Text handling and translations</h2>
- * Methods from <a href="https://fabric.moddedmc.wiki/text-and-translations">Text handling and translations</a> page.<br>
+ * Methods from <a href="https://fabric.moddedmc.wiki/text-and-translations">Text handling and translations</a> and
+ * <a href="https://fabric.moddedmc.wiki/data-generation/translations#programmatic-translation-example">Translation Provider</a> page.<br>
  */
-@SuppressWarnings({"UnnecessaryLocalVariable", "UnusedReturnValue"})
+@SuppressWarnings({"UnnecessaryLocalVariable", "UnusedReturnValue", "unused", "UnusedAssignment"})
 public class TextUtil {
     private TextUtil() {
     }
@@ -16,6 +21,7 @@ public class TextUtil {
     public static void initializeAllTextFunctions() {
         textLiterals();
         formattingText();
+        generateHumanReadable(new Identifier(MyMod.MOD_ID, "big_blue_boat"));
 
         LoggerUtil.devLogger("Used all Text functions");
     }
@@ -55,5 +61,12 @@ public class TextUtil {
     private static void formattingText() {
         MutableText result = deserializingTextFromJson();
         result = Text.literal("Hello World!").formatted(Formatting.AQUA, Formatting.BOLD, Formatting.UNDERLINE);
+    }
+
+    public static String generateHumanReadable(Identifier identifier) {
+        String identifier_path = identifier.getPath();
+        String lowercase = identifier_path.replace("_", " ");
+        String capitalized = WordUtils.capitalize(lowercase);
+        return capitalized;
     }
 }
